Refractoring homePage
